evaluation/requirements.txt
certifi==2020.12.5
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
tifffile==2021.4.8
tqdm==4.60.0
